[FLINK-3874] Minor refactoring
