+++ flink-streaming-connectors/flink-connector-kafka-0.8/src/main/java/org/apache/flink/streaming/connectors/kafka/Kafka08JsonTableSink.java
@@ -27,7 +27,7 @@ import java.util.Properties;
 /**
  * Kafka 0.8 {@link KafkaTableSink} that serializes data in JSON format.
  */
-public class Kafka08JsonTableSink extends KafkaJsonTableSinkBase {
+public class Kafka08JsonTableSink extends KafkaJsonTableSink {
 	/**
 	 * Creates {@link KafkaTableSink} for Kafka 0.8
 	 *
